+++ src/store/modules/user.js
@@ -1,13 +1,12 @@
 const CHANGE_USER_INFO = 'user/CHANGE_USER_INFO';
 
 export const changeUserInfo = (email, name, profileUrl) => {
-  return ({
-  type: CHANGE_USER_INFO,
-  email,
-  name,
-  profileUrl,
-})
-
+  return {
+    type: CHANGE_USER_INFO,
+    email,
+    name,
+    profileUrl,
+  };
 };
 
 const initialState = { email: '', name: '', profileUrl: '' };
